+++ lib/pages/cacheImageTablePage.dart
@@ -33,16 +33,7 @@ class _CacheImageTablePageState extends State<CacheImageTablePage> {
       body: ListView.builder(
         itemCount: list.length,
         itemBuilder: (ctx, index) => GestureDetector(
-          onTap: () {
-            // showDialog(
-            //   context: context,
-            //   builder: (ctx) => SimpleDialog(
-            //     children: <Widget>[
-            //       LocalCacheImage('https://segmentfault.com/img/bVbHoL3'),
-            //     ],
-            //   ),
-            // );
-          },
+          onTap: () {},
           child: Container(
             padding: EdgeInsets.symmetric(horizontal: 16, vertical: 8),
             child: Image.memory(